Extract config and getRenderedAttributeValue to separate module
We'll use these to create a snapshot test
